requirements.txt
streamlit
numpy
pandas
scipy
matplotlib
python-dotenv
pydicom
mne
nibabel
soundfile
